+++ CHANGELOG.md
@@ -24,6 +24,7 @@ metric events are swallowed for the time being.
 - Add event filtering to extensions.
 - Proper 404 page for web UI.
 - Add sensuctl extension command.
+- Add extensions to pipelined.
 
 ### Changed
 - Changed the maximum number of open file descriptors on a system to from 1024